Benchmarking/requirements.txt
# Dépendances pour le benchmarking des modèles YOLO
# Version finale avec support TFLite + ONNX

# Core dependencies
pyyaml>=6.0
opencv-python>=4.5.0
numpy>=1.21.0
tqdm>=4.62.0
psutil>=5.8.0

# Analysis and visualization
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# TFLite runtime
tflite-runtime>=2.5.0

# ONNX Runtime (avec support XNNPACK pour ARM)
onnxruntime>=1.10.0

# ROS2 dependencies (déjà installées avec ROS2)
# rclpy
# sensor_msgs
# cv_bridge

# Pour Coral TPU (Raspberry Pi uniquement, installer manuellement)
# Suivre les instructions: https://coral.ai/docs/accelerator/get-started/
# pycoral
# tflite-runtime (version Coral)
